✨ workgin docker build

+++ src/Command/BackgroundCleanupOldDeploymentsCommand.php
@@ -2,7 +2,6 @@
 
 namespace App\Command;
 
-use App\Dto\Deployment;
 use App\Dto\Size;
 use App\Service\FileSizeService;
 use App\Service\ProjectService;
@@ -11,12 +10,11 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
+use function array_keys;
 use function array_shift;
-use function assert;
 
 #[AsCommand(
     name: 'background:cleanup-old-deployments',
@@ -36,8 +34,15 @@ final class BackgroundCleanupOldDeploymentsCommand extends Command
         $this->maxBytes = $this->fileSizeService->convertToBytes($rahMaxDiskUsage);
     }
 
+    protected function configure(): void
+    {
+        $this->addOption('dry-run', 'd', null, 'Do not delete anything, just show what would be deleted');
+    }
+
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $output->writeln('<fg=white;bg=magenta>RAH Background Cleanup Old Deployments:</>');
+
         $io = new SymfonyStyle($input, $output);
         $this->requestStack->push(new Request());
         $projects = $this->projectService->loadAll();
@@ -55,16 +60,41 @@ final class BackgroundCleanupOldDeploymentsCommand extends Command
 
         ksort($deployments, SORT_NATURAL);
 
-        $io->text('before: ' . Size::formatSize($totalSize) . ' allowed: ' . Size::formatSize($this->maxBytes));
+        if ($totalSize <= $this->maxBytes) {
+            $message = 'No cleanup needed. Total size: ' . Size::formatSize($totalSize) . ' / ' . Size::formatSize($this->maxBytes);
+            $output->writeln('<fg=green>' . $message . '</>');
+            return Command::SUCCESS;
+        }
+
+        $printRows = [];
+        foreach ($deployments as $deployment) {
+            $printRows[] = [
+                'project' => $deployment->project->name,
+                'name' => $deployment->name,
+                'default' => $deployment->deploymentSettings->defaultDeployment === $deployment->name,
+                'size' => $deployment->size->humanReadable,
+                'lastUpdate' => $deployment->lastUpdate->format('Y-m-d H:i:s'),
+            ];
+        }
+
+        if ($printRows) {
+            $io->table(array_keys($printRows[0]), $printRows);
+        }
+
+        $output->writeln('before: ' . Size::formatSize($totalSize) . ' allowed: ' . Size::formatSize($this->maxBytes));
 
         while ($totalSize > $this->maxBytes && $deployments) {
             $deploymentToDelete = array_shift($deployments);
-            $io->caution('Deleting deployment: ' . $deploymentToDelete->project->name . '--' . $deploymentToDelete->name . ' freeing up ' . $deploymentToDelete->size->humanReadable);
-            $this->projectService->deleteDeployment($deploymentToDelete);
+            $io->caution('Deleting deployment: ' . $deploymentToDelete->project->name . ' ' . $deploymentToDelete->name . ' freeing up ' . $deploymentToDelete->size->humanReadable);
+            if (!$input->getOption('dry-run')) {
+                $this->projectService->deleteDeployment($deploymentToDelete);
+            }
+
             $totalSize -= $deploymentToDelete->size->bytes;
         }
 
-        $io->success('DONE. Total size left: ' . Size::formatSize($totalSize) . ' / ' . Size::formatSize($this->maxBytes));
+        $message = 'DONE. Total size left: ' . Size::formatSize($totalSize) . ' / ' . Size::formatSize($this->maxBytes);
+        $output->writeln('<fg=green>' . $message . '</>');
 
         return Command::SUCCESS;
     }
